+++ crates/api/src/launchpad.rs
@@ -1,17 +1,21 @@
 use miette::IntoDiagnostic;
 use moon_cache::{CacheEngine, cache_item};
-use moon_common::{consts::CONFIG_DIRNAME, is_test_env};
+use moon_common::{consts::CONFIG_DIRNAME, is_ci, is_test_env};
 use moon_env::MoonEnvironment;
 use moon_env_var::GlobalEnvBag;
 use moon_time::now_millis;
 use semver::Version;
 use serde::{Deserialize, Serialize};
 use starbase_utils::{fs, json};
-use std::env::{self, consts};
+use std::collections::BTreeMap;
+use std::env::consts;
 use std::path::Path;
+use std::sync::{Arc, OnceLock};
 use std::time::Duration;
 use tracing::{debug, instrument};
 use uuid::Uuid;
+
+static INSTANCE: OnceLock<Arc<Launchpad>> = OnceLock::new();
 
 const ALERT_PAUSE_DURATION: Duration = Duration::from_secs(43200); // 12 hours
 
@@ -29,6 +33,11 @@ cache_item!(
         pub remote_version: Option<Version>,
     }
 );
+
+#[derive(Serialize)]
+pub struct ToolchainUsage {
+    pub toolchains: BTreeMap<String, String>,
+}
 
 fn load_or_create_anonymous_uid(id_path: &Path) -> miette::Result<String> {
     if id_path.exists() {
@@ -60,13 +69,43 @@ pub struct VersionCheck {
     pub update_available: bool,
 }
 
-pub struct Launchpad;
+pub struct Launchpad {
+    #[allow(dead_code)]
+    moon_env: Arc<MoonEnvironment>,
+    moon_version: String,
+    user_id: String,
+    repo_id: Option<String>,
+}
 
 impl Launchpad {
+    pub fn register(moon_env: Arc<MoonEnvironment>) -> miette::Result<()> {
+        let user_id = load_or_create_anonymous_uid(&moon_env.id_file)?;
+
+        let repo_id = fs::find_upwards(CONFIG_DIRNAME, &moon_env.working_dir)
+            .map(|dir| create_anonymous_rid(dir.parent().unwrap()));
+
+        let moon_version = GlobalEnvBag::instance()
+            .get("MOON_VERSION")
+            .unwrap_or_default();
+
+        let _ = INSTANCE.set(Arc::new(Self {
+            moon_env,
+            moon_version,
+            user_id,
+            repo_id,
+        }));
+
+        Ok(())
+    }
+
+    pub fn instance() -> Arc<Launchpad> {
+        Arc::clone(INSTANCE.get().unwrap())
+    }
+
     #[instrument(skip_all)]
     pub async fn check_version(
+        &self,
         cache_engine: &CacheEngine,
-        moon_env: &MoonEnvironment,
         bypass_cache: bool,
         manifest_url: &str,
     ) -> miette::Result<Option<VersionCheck>> {
@@ -81,7 +120,7 @@ impl Launchpad {
             }
         }
 
-        if let Some(result) = Self::check_version_without_cache(moon_env, manifest_url).await? {
+        if let Some(result) = self.check_version_without_cache(manifest_url).await? {
             state.data.last_check_time = Some(now);
             state.data.local_version = Some(result.local_version.clone());
             state.data.remote_version = Some(result.remote_version.clone());
@@ -94,16 +133,14 @@ impl Launchpad {
     }
 
     pub async fn check_version_without_cache(
-        moon_env: &MoonEnvironment,
+        &self,
         manifest_url: &str,
     ) -> miette::Result<Option<VersionCheck>> {
         if is_test_env() || proto_core::is_offline() {
             return Ok(None);
         }
 
-        let version = GlobalEnvBag::instance()
-            .get("MOON_VERSION")
-            .unwrap_or_default();
+        let version = &self.moon_version;
 
         debug!(
             current_version = &version,
@@ -111,37 +148,18 @@ impl Launchpad {
             "Checking for a new version of moon"
         );
 
-        let mut client = reqwest::Client::new()
-            .get(manifest_url)
-            .header("X-Moon-OS", consts::OS.to_owned())
-            .header("X-Moon-Arch", consts::ARCH.to_owned())
-            .header("X-Moon-Version", &version)
-            .header("X-Moon-CI", ci_env::is_ci().to_string())
+        let request = self
+            .create_request(manifest_url)?
             .header(
                 "X-Moon-CI-Provider",
                 format!("{:?}", ci_env::detect_provider()),
             )
-            .header("X-Moon-CD", cd_env::is_cd().to_string())
             .header(
                 "X-Moon-CD-Provider",
                 format!("{:?}", cd_env::detect_provider()),
-            )
-            .header(
-                "X-Moon-UID",
-                load_or_create_anonymous_uid(&moon_env.id_file)?,
             );
 
-        if let Some(moon_dir) = fs::find_upwards(
-            CONFIG_DIRNAME,
-            env::current_dir().expect("Invalid working directory!"),
-        ) {
-            client = client.header(
-                "X-Moon-RID",
-                create_anonymous_rid(moon_dir.parent().unwrap()),
-            );
-        }
-
-        let Ok(response) = client.send().await else {
+        let Ok(response) = request.send().await else {
             return Ok(None);
         };
 
@@ -150,7 +168,7 @@ impl Launchpad {
         };
 
         let data: CurrentVersion = json::parse(text)?;
-        let local_version = Version::parse(&version).into_diagnostic()?;
+        let local_version = Version::parse(version).into_diagnostic()?;
         let remote_version = Version::parse(&data.current_version).into_diagnostic()?;
         let update_available = remote_version > local_version;
 
@@ -167,5 +185,39 @@ impl Launchpad {
             message: data.message,
             update_available,
         }))
+    }
+
+    pub async fn track_toolchain_usage(
+        &self,
+        toolchains: BTreeMap<String, String>,
+    ) -> miette::Result<()> {
+        if !is_ci() || is_test_env() || proto_core::is_offline() {
+            return Ok(());
+        }
+
+        let request = self
+            .create_request("https://launch.moonrepo.app/moon/toolchain_usage")?
+            .json(&ToolchainUsage { toolchains });
+
+        let _response = request.send().await.into_diagnostic()?;
+
+        Ok(())
+    }
+
+    fn create_request(&self, url: &str) -> miette::Result<reqwest::RequestBuilder> {
+        let mut client = reqwest::Client::new()
+            .post(url)
+            .header("X-Moon-OS", consts::OS.to_owned())
+            .header("X-Moon-Arch", consts::ARCH.to_owned())
+            .header("X-Moon-Version", self.moon_version.clone())
+            .header("X-Moon-CI", ci_env::is_ci().to_string())
+            .header("X-Moon-CD", cd_env::is_cd().to_string())
+            .header("X-Moon-UID", self.user_id.clone());
+
+        if let Some(repo_id) = &self.repo_id {
+            client = client.header("X-Moon-RID", repo_id.to_owned());
+        }
+
+        Ok(client)
     }
 }