+++ crates/app/src/systems/execute.rs
@@ -2,14 +2,12 @@ use moon_api::Launchpad;
 use moon_cache::CacheEngine;
 use moon_common::{color, is_formatted_output, is_test_env};
 use moon_console::{Checkpoint, Console};
-use moon_env::MoonEnvironment;
 use starbase::AppResult;
 use tracing::{debug, instrument};
 
 #[instrument(skip_all)]
 pub async fn check_for_new_version(
     console: &Console,
-    moon_env: &MoonEnvironment,
     cache_engine: &CacheEngine,
     manifest_url: &str,
 ) -> AppResult {
@@ -17,7 +15,10 @@ pub async fn check_for_new_version(
         return Ok(None);
     }
 
-    match Launchpad::check_version(cache_engine, moon_env, false, manifest_url).await {
+    match Launchpad::instance()
+        .check_version(cache_engine, false, manifest_url)
+        .await
+    {
         Ok(Some(result)) => {
             if !result.update_available {
                 return Ok(None);